+++ tests/settings/manage_script.py
@@ -1,3 +1,3 @@
 from .base import *  # noqa: F403
 
-DJANGO_MANAGE_SCRIPT = "mycli"
+DT_MANAGE_SCRIPT = "mycli"